+++ js/main.js
@@ -1,6 +1,4 @@
-import { getFetchDataAndRender } from './thumbnails.js';
 import './api.js';
 import './uploadForm.js';
 import './slider.js';
-
-getFetchDataAndRender();
+import './filters.js';
