+++ js/thumbnails.js
@@ -1,11 +1,9 @@
-import { getData } from './api.js';
 import { openBigPicture } from './fullSize.js';
 
 const templateUserPicture = document.querySelector('#picture')
   .content
   .querySelector('.picture');
 
-let usersPictures = [];
 
 const containerUsersPictures = document.querySelector('.pictures');
 
@@ -21,7 +19,7 @@ const renderUsersPictures = (pictures) => {
     userPicture.dataset.id = id;
     usersPicturesFragment.append(userPicture);
     userPicture.addEventListener('click', (event) => {
-      const currentPicture = usersPictures.find((photo) => event.currentTarget.dataset.id === photo.id.toString());
+      const currentPicture = pictures.find((photo) => event.currentTarget.dataset.id === photo.id.toString());
       openBigPicture(currentPicture);
     });
   });
@@ -32,17 +30,5 @@ const clearUsersPictures = () => {
   containerUsersPictures.innerHTML = '';
 };
 
-const getFetchDataAndRender = () => {
-  getData().then((data) => {
-    usersPictures = data;
-    renderUsersPictures(usersPictures);
-  }).catch(() => {
-    const errorTemplate = document.querySelector('#data-error').content.cloneNode(true);
-    document.body.appendChild(errorTemplate);
-    setTimeout(() => {
-      document.body.removeChild(document.querySelector('.data-error'));
-    }, 5000);
-  });
-};
 
-export { renderUsersPictures, clearUsersPictures, getFetchDataAndRender };
+export { renderUsersPictures, clearUsersPictures };
